Added Castamere Hero
Basically just added my hero to the project nothing no major. I also
added the lines of code needed to connect him to the cam as well as
moving him relative the surfaces height. I was not able to get the
orientation relative to the surface to work correctly. Also there are a
number of but with his coloring and his animation that I may need help
figuring out.