Fps displayed and viewport mostly working.
This is a huge commit (sorry about that) I got fps to display to the
bottom right side of the window. I also got a picture in picture
viewport to display in the bottom left. In addition to these things I
am still working on getting the world class up and running, it has much
more functionality then it has in the past.